Support XDG directories
This searches XDG_CONFIG_HOME and XDG_CONFIG_DIRS for configuration, and
XDG_DATA_HOME and XDG_DATA_DIRS for data files. For configuration, the
first found config file is used by itself, but we use all data files
that we find in any location. This will hopefully help with packaging
splatmoji.