Use ~ to refer to HOME
The history file will now default to ~/.local/state/splatmoji/history as
per XDG proposal for such things. The tilde is expanded to ${HOME} when
reading the config. This expansion will occur for every variable, as
long as the tilde occurs at the start of the variable value. I don't see
this causing any issues, and could be useful if you have either menu or
typing tools in ~/bin/